Dealer Mainly
I worked on my dealer, and being able to print out some cards from the
deck. Also I was able to print out if the person choose to hit. After
that I need to be able to calculate the value of the persons hand to
see if they bust, this will also help my dealer who will just play
until they are above 17.